src/AiAnalyst/requirements.txt
finnhub-python
requests
chromadb
ollama